Always end long path with separator
Fix bug in __sd_extract_next_part by always ending the returned long
path with the separator.

This is necessary because the short path will always end with a
separator, and it's possible for the final path to end with either the
short path followed by the basename, or the short path followed by the
long path followed by the basename.