Handle when the path is only a basename
Make the same kind of bugfix in __sd_extract_last_part that was
previously made in __sd_extract_next_part. Namely, handle the case where
the path does not contain the separator.

We make the part the entire original long_path, and consider the rest of
the string to be empty as a special case, because we don't want to add a
separator where there wasn't previously one.